+++ source/app/service-providers/windows/create-print-window.ts
@@ -19,7 +19,6 @@ import {
   type BrowserWindowConstructorOptions
 } from 'electron'
 import attachLogger from './attach-logger'
-import preventNavigation from './prevent-navigation'
 import setWindowChrome from './set-window-chrome'
 import type { WindowPosition } from './types'
 
@@ -65,9 +64,6 @@ export default function createPrintWindow (logger: LogProvider, config: ConfigPr
     })
 
   // EVENT LISTENERS
-
-  // Prevent arbitrary navigation away from our WEBPACK_ENTRY
-  preventNavigation(logger, window)
 
   // Implement main process logging
   attachLogger(logger, window, 'Print Window')